backend/requirements.txt
opencv-python        # phát hiện khuôn mặt
Pillow               # xử lí ảnh
sqlalchemy           # ORM
pymysql
typing-extensions
python-multipart     # upload ảnh
ultralytics
mysql-connector-python